If there is a label on the camera show it
For many cameras it is difficult to tell what camera you want to look at.  This change updates foggycam.py to create the directory with the label name if there is one set up for the camera. 

FYI I am a c#/web dev, and this is one of my first attempts using python. Any advice appreciated.